fix company transaction history

+++ ccc_webapp/src/main/java/Utils_db/CitizenTransaction.java
@@ -67,14 +67,14 @@ public class CitizenTransaction {
     /**
      * @return the quantity
      */
-    public int getQuantity() {
+    public double getQuantity() {
         return quantity;
     }
 
     /**
      * @param quantity the quantity to set
      */
-    public void setQuantity(int quantity) {
+    public void setQuantity(double quantity) {
         this.quantity = quantity;
     }
